- Removed console output - Fixed crash on comparing multiple enrollments - Provisional fix for incorrect plan year bug - Further refactoring
